started code to prepare mustard data; renamed data_prep.py prepare_data.py
