+++ models_test/asist_test.py
@@ -9,7 +9,7 @@ from models.baselines import LRBaseline, EmbeddingsOnly
 from models.train_and_test_models import *
 from models.input_models import *
 
-from data_prep.data_prep import *
+from data_prep.prepare_data import *
 import data_prep.sentiment_score_prep as score_prep
 import data_prep.asist_data.asist_prep as asist_prep
