+++ models_test/model_test.py
@@ -6,7 +6,7 @@ from models.baselines import LRBaseline, EmbeddingsOnly
 from models.train_and_test_models import *
 
 from models.input_models import *
-from data_prep.data_prep import *
+from data_prep.prepare_data import *
 
 # import parameters for model
 # comment or uncomment as needed
